implement HttpController component
