+++ client/src/components/Page.svelte
@@ -1,0 +1,7 @@
+<script>
+    import HttpController from "./HttpController.svelte";
+</script>
+
+<HttpController>
+    <slot></slot>
+</HttpController>
